refactor(models): DisclosureReport is the submitted report, move to models.h1
Every DisclosureReport field maps directly onto the POST /reports payload
built in tools/h1_api.H1Client.submit_report - title -> attributes.title,
body_markdown -> attributes.vulnerability_information, impact_statement ->
attributes.impact, weakness_id -> attributes.weakness_id, programme_handle ->
relationships.program.data.id, severity -> attributes.severity_rating. The
class IS the H1 wire shape, not an intermediate authored by the Technical
Author that gets translated at the boundary - the docstring already said
"in H1 format". Belongs in models.h1 alongside SubmissionResult.

To make h1.py's `from models import VerifiedVulnerability` resolvable at
import time, the vulnerability-findings section (RawFinding,
VerifiedVulnerability) moves above the `from models.h1 import ...` line in
models/__init__.py. Declaration order is now slightly out of pipeline order
(findings declared before recon) - documented in the module docstring so
the next reader does not "tidy" it back into pipeline order and break the
import.

Five importers migrate to "from models.h1 import DisclosureReport"; no
back-compat re-export.
