+++ tests/conftest.py
@@ -16,14 +16,18 @@ os.environ.setdefault("CYBERSQUAD_CONTACT_EMAIL", "ci@example.invalid")
 import pytest  # noqa: E402
 
 from models import (  # noqa: E402
-    DisclosureReport,
     Endpoint,
     RawFinding,
     ReconResult,
     Severity,
     VerifiedVulnerability,
 )
-from models.h1 import Programme, ScopeItem, ScopeType  # noqa: E402
+from models.h1 import (  # noqa: E402
+    DisclosureReport,
+    Programme,
+    ScopeItem,
+    ScopeType,
+)
 
 
 # Probe tools call tools._helpers.adaptive_sleep between requests for rate-limit
